chore: complete manual script hygiene phase
